requirements.txt
anyio==4.2.0
asgiref==3.7.2
cfgv==3.4.0
channels==4.0.0
channels-redis==4.2.0
click==8.1.7
distlib==0.3.8
Django==3.2.4
django-cors-headers==4.3.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-querycount==0.8.3
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
h11==0.14.0
httptools==0.6.1
identify==2.5.33
idna==3.6
inflection==0.5.1
msgpack==1.0.7
nodeenv==1.8.0
packaging==23.2
Pillow==10.0.0
platformdirs==4.2.0
pre-commit==3.6.0
#psycopg2==2.9.6
#psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.1
python-environ==0.4.54
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
setuptools==69.0.3
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.27.0.post1
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
